Join logical domains in whole module
