+++ backend/yagram-core/src/domains/mod.rs
@@ -1,0 +1,2 @@
+pub mod auth;
+pub mod users;
