+++ backend/yagram-core/src/domains/users/mod.rs
@@ -1,0 +1,1 @@
+pub mod handlers;
